feat: 完善用户角色cuda

+++ app/http/controller/response.go
@@ -6,7 +6,7 @@ type Response struct {
 	Msg         string      `json:"message,omitempty"`      // 提示
 	Type        string      `json:"type,omitempty"`         // 类型
 	Data        interface{} `json:"result,omitempty"`       // 数据
-	Total       int         `json:"total,omitempty"`        // 总条数
+	Total       int64       `json:"total,omitempty"`        // 总条数
 	CurrentPage int         `json:"current_page,omitempty"` // 当前页
 	TotalPage   int         `json:"total_page,omitempty"`   // 总页数
 }
